requirements.txt
numpy
pyinstaller
python-dotenv
